+++ lib/use-agent-memo.ts
@@ -1,5 +1,6 @@
 import { useContext, useEffect, useState } from "react";
 import { AgentContextInternal, implicitRootAgentContext } from "./agent-context";
+import { getPathKey } from "./get-path-key";
 
 export function useAgentMemo<T>(
   name: string,
@@ -13,15 +14,24 @@ export function useAgentMemo<T>(
 ): T {
   const [latestValue, setLatestValue] = useState<T>(factory());
   const context = useContext(AgentContextInternal);
+  const { prefix, path } = getPathKey(context.breadcrumbs, name);
 
   useEffect(() => {
-    if (options?.enabled === false) return void implicitRootAgentContext.delete(name);
+    if (options?.enabled === false) return void implicitRootAgentContext.delete(path);
 
     const newValue = factory();
-    implicitRootAgentContext.set(name, { type: "state", data: newValue, context, description: options?.description });
+    implicitRootAgentContext.set(path, {
+      name,
+      prefix,
+      type: "state",
+      data: newValue,
+      context,
+      description: options?.description,
+    });
     setLatestValue(newValue);
-    return () => void implicitRootAgentContext.delete(name);
-  }, [name, ...dependencies, options?.enabled, options?.description]);
+
+    return () => void implicitRootAgentContext.delete(path);
+  }, [path, prefix, name, options?.enabled, options?.description, ...dependencies]);
 
   return latestValue;
 }